resolved god file
