+++ src/api/db/utility/project-utility.ts
@@ -1,231 +1,8 @@
+import {Database} from "../db";
+import {ValProject, ValUser} from "../validation";
+import {Ability, Like, Project, ProjectMember, View} from "../../../models";
 
-import {Database} from "./db";
-import {checkInvalidChars, ValMessage, ValNotification, ValProject, ValUser} from "./validation";
-import {Ability, DirectChat, Like, Message, Notification, Project, ProjectMember, User, View} from "../../models";
-
-export class Utility {
-
-    /* region User */
-
-    /* region Base */
-    static async addUser(userId: string, username: string, firstname: string, lastname: string, pfp: string, email:string, clazz:string,
-                   birthdate: Date, biografie: string, permissions: number, department: string): Promise<boolean> {
-        try {
-            const id = userId.toLowerCase();
-
-            if (!ValUser.isValid(id, username, firstname, lastname, email, clazz, birthdate, biografie, permissions, department)) {
-                return false;
-            }
-
-            return await Database.addUser(id, username, firstname, lastname, pfp, email, clazz, birthdate.toDateString(), biografie, permissions, department);
-        }
-        catch (e) {
-            return false;
-        }
-    }
-
-    static async getUser(userId: string): Promise<User | null> {
-        try {
-            const id = userId.toLowerCase();
-
-            if(!ValUser.isUserIdValid(id)) {
-                return null;
-            }
-
-            const user = await Database.getUser(id);
-            if (!user) {
-                return null;
-            }
-            return user;
-        } catch (e) {
-            return null;
-        }
-    }
-    
-    static async getUserIdByFullName(fullName: string): Promise<string | null> {
-        try {
-            const names = fullName.split("-");
-
-            if(fullName.length < 1 || fullName.length > 40 || names.length !== 2 || names[0].length < 1 || names[1].length < 1 || !checkInvalidChars(names[0]) || !checkInvalidChars(names[1])) {
-                return null;
-            }
-
-            //first letter uppercase and the rest lowercase
-            const firstname = names[0].charAt(0).toUpperCase() + names[0].slice(1).toLowerCase();
-
-            //first letter uppercase and the rest lowercase
-            const lastname = names[1].charAt(0).toUpperCase() + names[1].slice(1).toLowerCase();
-
-            return await Database.getUserIdByFullName(firstname, lastname);
-        }
-        catch (e) {
-            return null;
-        }
-    }
-
-    static async updateUser(userId: string, username: string, firstname: string, lastname: string, pfp:string, email:string, clazz:string, birthdate: Date, biografie: string, permissions: number, department: string): Promise<boolean> {
-        try {
-            const id = userId.toLowerCase();
-
-            if(!ValUser.isValid(id, username, firstname, lastname, email, clazz, birthdate, biografie, permissions, department)) {
-                return false;
-            }
-
-            return await Database.updateUser(id, username, firstname, lastname, pfp, email, clazz, birthdate.toDateString(), biografie, permissions, department);
-        }
-        catch (e) {
-            return false;
-        }
-    }
-
-    static async deleteUser(userId: string): Promise<boolean> {
-        try {
-            const id = userId.toLowerCase();
-
-            if(!ValUser.isUserIdValid(id)) {
-                return false;
-            }
-
-            return await Database.deleteUser(id);
-        }
-        catch (e) {
-            return false;
-        }
-    }
-
-    static async getUsers(): Promise<User[] | null> {
-        try {
-            return await Database.getUsers();
-        }
-        catch (e) {
-            return null;
-        }
-    }
-
-    /* endregion */
-
-    /* region UserAbility */
-    static async addUserAbility(userId: string, abilityId: number): Promise<boolean> {
-        try {
-            const id = userId.toLowerCase();
-
-            if(!await ValUser.isUserValid(id) || abilityId < 1) {
-                return false;
-            }
-
-            if(await Database.userAbilityAlreadyExists(id, abilityId)) {
-                return false;
-            }
-
-            return await Database.addUserAbility(id, abilityId);
-        }
-        catch (e) {
-            return false;
-        }
-    }
-
-    static async getUserAbilities(userId: string): Promise<Ability[] | null> {
-        try {
-            const id = userId.toLowerCase();
-
-            if(!await ValUser.isUserValid(id)) {
-                return null;
-            }
-
-            return await Database.getUserAbilitiesByUserId(id);
-        }
-        catch (e) {
-            return null;
-        }
-    }
-
-    static async deleteUserAbility(userId: string, abilityId: number): Promise<boolean> {
-        try {
-            const id = userId.toLowerCase();
-
-            return await Database.deleteUserAbility(id, abilityId);
-        }
-        catch (e) {
-            return false;
-        }
-    }
-
-    /* endregion */
-
-    /* region Notification */
-
-    static async addNotification(userId: string, title: string, text: string): Promise<boolean> {
-        try {
-            const id = userId.toLowerCase();
-
-            if(!ValNotification.isValid(id, title, text)) {
-                return false;
-            }
-
-            if(!await ValUser.isUserValid(id)) {
-                return false;
-            }
-
-            const date = new Date(Date.now());
-
-            const dateTime = `${date.toLocaleDateString()};${date.toLocaleTimeString()}`;
-
-            return await Database.addNotification(id, title, text, dateTime);
-        }
-        catch (e) {
-            return false;
-        }
-    }
-
-    static async getNotifications(userId: string): Promise<Notification[] | null> {
-        try {
-            const id = userId.toLowerCase();
-
-            if(!ValUser.isUserIdValid(id)) {
-                return null;
-            }
-
-            return await Database.getNotificationsByUserId(id);
-        }
-        catch (e) {
-            return null;
-        }
-    }
-
-    static async deleteNotification(userId: string, notificationId: number): Promise<boolean> {
-        try {
-            const id = userId.toLowerCase();
-
-            if(!await this.isNotificationOwner(id, notificationId)) {
-                return false;
-            }
-
-            return await Database.deleteNotification(notificationId);
-        }
-        catch (e) {
-            return false;
-        }
-    }
-
-    static async isNotificationOwner(userId: string, notificationId: number): Promise<boolean> {
-        try {
-            const id = userId.toLowerCase();
-
-            if(!ValUser.isUserIdValid(id) || notificationId < 1) {
-                return false;
-            }
-
-            return await Database.isNotificationOwner(id, notificationId);
-        }
-        catch (e) {
-            return false;
-        }
-    }
-
-    /* endregion */
-
-    /* region Project */
-
+export class ProjectUtility {
     /* region Base */
     static async addProject(name: string, ownerId: string, thumbnail: string, description: string, links: string, maxMembers: number): Promise<boolean> {
         try {
@@ -654,15 +431,6 @@ export class Utility {
         }
     }
 
-    static async getAllAbilities(): Promise<Ability[]> {
-        try {
-            return await Database.getAbilities();
-        }
-        catch (e) {
-            return null;
-        }
-    }
-
     static async deleteAbilityFromProject(projectId: number, abilityId: number): Promise<boolean> {
         try {
             return await Database.deleteProjectAbility(projectId, abilityId);
@@ -671,169 +439,5 @@ export class Utility {
             return false;
         }
     }
-    /* endregion */
-
-    /* endregion */
-
-    /* region DirectChat */
-
-    /* region Chat */
-
-    static async addDirectChat(userId: string, otherUserId: string): Promise<boolean> {
-        try {
-            const id = userId.toLowerCase();
-            const otherId = otherUserId.toLowerCase();
-
-            if(!await ValUser.isUserValid(id) || !await ValUser.isUserValid(otherId) || id === otherId) {
-                return false;
-            }
-
-            const chat = await Database.getDirectChatByUserIds(id, otherId);
-
-            if(chat !== null) {
-                return false;
-            }
-
-            return await Database.addDirectChat(id, otherId);
-        }
-        catch (e) {
-            return false;
-        }
-    }
-
-    static async getDirectChats(userId: string): Promise<DirectChat[]> {
-        try {
-            const id = userId.toLowerCase();
-
-            if(!ValUser.isUserIdValid(id)) {
-                return null;
-            }
-
-            return await Database.getDirectChatsByUserId(id);
-        }
-        catch (e) {
-            return null;
-        }
-    }
-
-    static async getDirectChat(userId: string, otherUserId: string): Promise<DirectChat | null> {
-        try {
-            const id = userId.toLowerCase();
-            const otherId = otherUserId.toLowerCase();
-
-            if(!ValUser.isUserIdValid(id) || !ValUser.isUserIdValid(otherId) || id === otherId) {
-                return null;
-            }
-
-            return await Database.getDirectChatByUserIds(id, otherId);
-        }
-        catch (e) {
-            return null;
-        }
-    }
-
-    static async deleteDirectChat(userId: string, otherUserId: string): Promise<boolean> {
-        try {
-            const id = userId.toLowerCase();
-            const otherId = otherUserId.toLowerCase();
-
-            if(!ValUser.isUserIdValid(id) || !ValUser.isUserIdValid(otherId) || id === otherId) {
-                return false;
-            }
-
-            const chat = await Database.getDirectChatByUserIds(id, otherId);
-
-            if(chat === null || chat.userId !== id || chat.otherUserId !== otherId) {
-                return false;
-            }
-
-            return await Database.deleteDirectChat(chat.id);
-        }
-        catch (e) {
-            return false;
-        }
-    }
-
-    /* endregion */
-
-    /* region Message */
-
-    static async addMessage(chatId: number, userId: string, message: string): Promise<boolean> {
-        try {
-            const id = userId.toLowerCase();
-
-            if(!ValMessage.isValid(id, message) || chatId < 1){
-                return false;
-            }
-
-            const chat = await Database.getDirectChatById(chatId);
-
-            if(chat === null || !(chat.userId === id || chat.otherUserId === id)) {
-                return false;
-            }
-
-            const date = new Date(Date.now());
-
-            const dateTime = `${date.toLocaleDateString()};${date.toLocaleTimeString()}`;
-
-            return await Database.addMessage(chatId, id, message, dateTime);
-        }
-        catch (e) {
-            return false;
-        }
-    }
-
-    static async getMessages(chatId: number): Promise<Message[]> {
-        try {
-            return await Database.getMessagesByChatId(chatId);
-        }
-        catch (e) {
-            return null;
-        }
-    }
-
-    static async updateMessage(messageId: number, chatId: number, userId: string, message: string): Promise<boolean> {
-        try {
-            const id = userId.toLowerCase();
-
-            if(!ValMessage.isValid(id, message) || messageId < 1 || chatId < 1){
-                return false;
-            }
-
-            return await Database.updateMessage(messageId, chatId, id, message);
-        }
-        catch (e) {
-            return false;
-        }
-    }
-
-    static async deleteMessage(userId: string, messageId: number): Promise<boolean> {
-        try {
-            const id = userId.toLowerCase();
-
-            if(!ValUser.isUserIdValid(id) || messageId < 1) {
-                return false;
-            }
-
-            return await Database.deleteMessage(messageId);
-        }
-        catch (e) {
-            return false;
-        }
-    }
-
-    /* endregion */
-
-    /* endregion */
-
-    /* region Ability */
-        static async addAbility(name: string, parentId: number | null): Promise<boolean> {
-            try {
-                return await Database.addAbility(name, parentId);
-            } catch (e) {
-                return false;
-            }
-        }
-
     /* endregion */
 }
